requirements.txt
PyYaml >= 6.0, <7.0

pii-data >= 0.1.0, <1.0.0
python-docx >= 0.8.11, <1.0.0
